Move each port type to a new header
