+++ VERSION
@@ -1,1 +1,1 @@
-0.4.677
+0.4.678
