+++ src/kernel/standard_input_port.cpp
@@ -14,31 +14,25 @@
    limitations under the License.
 */
 
-#include <meevax/kernel/string.hpp>
-#include <meevax/kernel/string_port.hpp>
+#include <meevax/kernel/standard_input_port.hpp>
 
 namespace meevax
 {
 inline namespace kernel
 {
-  string_port::operator std::istream &()
+  standard_input_port::operator std::istream &()
   {
-    return stringstream;
+    return std::cin;
   }
 
-  string_port::operator std::istream const&() const
+  standard_input_port::operator std::istream const&() const
   {
-    return stringstream;
+    return std::cin;
   }
 
-  string_port::operator std::ostream &()
+  auto operator <<(std::ostream & output, standard_input_port const&) -> std::ostream &
   {
-    return stringstream;
-  }
-
-  auto operator <<(std::ostream & output, string_port const& datum) -> std::ostream &
-  {
-    return output << magenta("#,(") << blue("open-string ") << string(datum.stringstream.str()) << magenta(")");
+    return output << magenta("#,(") << blue("standard-input") << magenta(")");
   }
 } // namespace kernel
 } // namespace meevax
